Move writing functionality to separate writing module
