+++ src/engine/strat_engine/backstore/backstore.rs
@@ -21,10 +21,10 @@ use crate::{
                 data_tier::DataTier,
                 metadata::MDADataSize,
             },
-            device::wipe_sectors,
             dm::get_dm,
             names::{format_backstore_ids, CacheRole, KeyDescription},
             serde_structs::{BackstoreSave, CapSave, Recordable},
+            writing::wipe_sectors,
         },
         BlockDevTier, DevUuid, PoolUuid,
     },
