+++ src/engine/strat_engine/backstore/metadata/bda.rs
@@ -14,7 +14,7 @@ use crate::{
                 sizes::{BDAExtendedSize, BlockdevSize, MDADataSize, STATIC_HEADER_SIZE},
                 static_header::{MetadataLocation, StaticHeader, StratisIdentifiers},
             },
-            device::SyncAll,
+            writing::SyncAll,
         },
         types::{DevUuid, PoolUuid},
     },
