+++ src/engine/strat_engine/backstore/metadata/static_header.rs
@@ -21,7 +21,7 @@ use crate::{
             backstore::metadata::sizes::{
                 static_header_size, BDAExtendedSize, BlockdevSize, MDASize, ReservedSize,
             },
-            device::SyncAll,
+            writing::SyncAll,
         },
         types::{DevUuid, PoolUuid},
     },
